refactor(main.go): refactored main.go by seperating to /internal/agent/agent.go and /internal/middleware/keycloak/keycloak.go
